pyproject.toml
[tool.poetry]
name = "brmanga"
version = "0.1.0"
description = ""
authors = ["gustavooliveira <gusatlan@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
selenium = "^4.14.0"
requests = "^2.31.0"
pillow = "^10.1.0"
pypdf2 = "^3.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
